requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2
djangorestframework==3.14.0
flake8==7.0.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.10.0
phonenumberslite==8.13.32
python-magic==0.4.27
psycopg2>=2.8.6,<2.9